Refactor action to support multiple clouds.
