+++ index.js
@@ -1,59 +1,29 @@
 const core = require('@actions/core');
-const { FieldValue, Firestore } = require('@google-cloud/firestore');
 const github = require('@actions/github');
-
-const db = new Firestore();
+const gcp = require('./gcp');
 
 async function run() {
   try {
+    const provider = core.getInput('cloud_provider');
     const table_name = core.getInput('table');
-    const table = db.collection(table_name);
     const number = github.context.issue.number;
     const repo = github.context.payload.repository.full_name;
     const pr = `${repo}/pr-${number}`
-    const branch = github.context.payload.pull_request.head.ref; 
-    let env;
+    const branch = github.context.payload.pull_request.head.ref;
+
     let data;
-
-    // First check if there is already an environment assigned to this PR.
-    core.info(`Checking if there is already an environment assinged to ${pr}...`);
-    const current_envs = await table.where('pull_requests', 'array-contains', pr).where('in_use', '==', true).limit(1).get();
-
-    if (!current_envs.empty) {
-      core.info(`Found an active QA environment for ${pr}`);
-      env = current_envs.docs[0];
-    } else {
-       // Check for another env using the same branch name
-      core.info(`No active QA environment for ${pr}. Looking for an active QA environment with the same branch (${branch})..`);
-      const branch_envs = await table.where('branch', '==', branch).where('in_use', '==', true).limit(1).get();
-      if (!branch_envs.empty) {
-        core.info(`Found an active QA environment for branch ${branch}. Adding ${pr} to it...`);
-        env = branch_envs.docs[0];
-        
-      } else {
-        // No environment currently assigned. Check for available environments.
-        core.info(`No active QA environment for ${pr}. Looking for an available environment...`);
-        const envs = await table.where('in_use', '==', false).limit(1).get();
-
-        // No environments available. Bail out.
-        if (envs.empty) {
-          throw new Error('No QA environments available.');
-        }
-    
-        env = envs.docs[0];
-      }
-      // Mark the environment as in_use and set the PR for reference.
-      core.info(`Found an available environment. Marking it as in_use by ${pr}`);
-      env.ref.update({
-        in_use: true,
-        branch: branch,
-        pull_requests: FieldValue.arrayUnion(pr),
-      });
+    switch (provider) {
+      case 'gcp':
+        data = await gcp(table_name, pr, branch);
+        break;
+      case 'aws':
+        throw new Error('AWS support is not yet implemented');
+      default:
+        throw new Error(`Unrecognized provider ${provider}. Only 'gcp' and 'aws' are supported.`);
     }
-
-    data = env.data();
     core.setOutput('url', data.url);
     core.setOutput('env_name', data.env_name)
+    core.notice(`PR ${pr} for branch ${branch} assigned to QA environment ${data.env_name} with url ${data.url}`);
   } catch (error) {
     core.setFailed(error.message);
   }
